requirements.txt
fastapi
uvicorn[standard]
streamlit
pandas
joblib
scikit-learn
pydantic
Pillow
requests
redis
